+++ robusta-dxf/src/parse/lwpolyline.rs
@@ -4,7 +4,7 @@ use crate::*;
 pub fn to_points(specific: &LwPolyline) -> Vec<Point> {
     let mut points = Vec::new();
     for vertex in &specific.vertices {
-        points.push(Point::new(vertex.x, vertex.y, 0.));
+        points.push(Point::new(vertex.x as f32, vertex.y as f32, 0.));
     }
 
     return points;
